+++ packages/moengage_cards/moengage_cards_ios/lib/moengage_cards_ios.dart
@@ -40,13 +40,11 @@ class MoEngageCardsIOS extends MoEngageCardsPlatform {
   }
 
   @override
-  void setAppOpenCardsSyncListener(
-    CardsSyncListener cardsSyncListener,
-    String appId,
-  ) {
-    super.setAppOpenCardsSyncListener(cardsSyncListener, appId);
+  void setSyncCompleteListener(
+      CardsSyncListener cardsSyncListener, String appId) {
+    super.setSyncCompleteListener(cardsSyncListener, appId);
     methodChannel.invokeMethod(
-      methodSetAppOpenCardsSyncListener,
+      methodSetCardsSyncListener,
       getAccountMeta(appId),
     );
   }