+++ packages/moengage_flutter/moengage_flutter/lib/src/moengage_flutter.dart
@@ -401,4 +401,18 @@ class MoEngageFlutter {
   void showNudge({MoEngageNudgePosition position = MoEngageNudgePosition.any}) {
     _platform.showNudge(position, appId);
   }
+
+  /// Get Multiple Self Handled InApps
+  /// @returns - Instance of [Future] of type [SelfHandledCampaignsData]
+  /// @since 8.1.0
+  Future<SelfHandledCampaignsData> getSelfHandledInApps() {
+    return _platform.getSelfHandledInApps(appId);
+  }
+
+  /// Register for Provisional Push Notification
+  /// Note: This API is only for iOS Platform.
+  /// @since 8.1.0
+  void registerForProvisionalPush() {
+    _platform.registerForProvisionalPush();
+  }
 }