+++ packages/moengage_flutter/moengage_flutter_android/android/src/main/kotlin/com/moengage/flutter/Constants.kt
@@ -49,3 +49,7 @@ const val ERROR_CODE_DELETE_USER = "DELETE_USER_ERROR"
 
 // Non Intrusive Nudge
 const val METHOD_NAME_SHOW_NUDGE = "showNudge"
+
+// Self Handled InApps
+const val METHOD_NAME_SELF_HANDLED_IN_APPS = "selfHandledInApps"
+const val ERROR_CODE_SELF_HANDLED_IN_APPS = "SELF_HANDLED_IN_APPS_ERROR"