+++ packages/moengage_flutter/moengage_flutter_ios/ios/Classes/MoEngageFlutterBridge.swift
@@ -1,6 +1,7 @@
 import Flutter
 import UIKit
 import MoEngagePluginBase
+import MoEngageCore
 
 public class MoEngageFlutterBridge: NSObject, FlutterPlugin {
     
@@ -17,13 +18,14 @@ public class MoEngageFlutterBridge: NSObject, FlutterPlugin {
         switch call.method {
         case MoEngageFlutterConstants.MethodNames.kRegisterForPush:
             MoEngagePluginBridge.sharedInstance.registerForPush()
-        
+        case MoEngageFlutterConstants.MethodNames.kRegisterForProvisionalPush:
+            registerForProvisionalPush()
         default:
-            handleWithPayload(call: call)
+            handleWithPayload(call: call, result: result)
         }
     }
     
-    private func handleWithPayload(call: FlutterMethodCall) {
+    private func handleWithPayload(call: FlutterMethodCall,  result: @escaping FlutterResult) {
         guard let payload = call.arguments as? [String: Any] else { return }
         switch call.method {
         case MoEngageFlutterConstants.MethodNames.kInitializeFlutter:
@@ -57,7 +59,10 @@ public class MoEngageFlutterBridge: NSObject, FlutterPlugin {
             MoEngagePluginBridge.sharedInstance.setAlias(payload)
         case MoEngageFlutterConstants.MethodNames.kResetUser:
             MoEngagePluginBridge.sharedInstance.resetUser(payload)
-            
+        case MoEngageFlutterConstants.MethodNames.kGetSelfHandleInApps:
+            MoEngagePluginBridge.sharedInstance.getSelfHandledInApps(payload) { campaignPayload in
+                MoEngageFlutterUtil.resume(channel: call.method, havingResult: result, withData: campaignPayload)
+            }
         default:
             print("Invalid invocation: \(call.method)")
         }
@@ -66,6 +71,15 @@ public class MoEngageFlutterBridge: NSObject, FlutterPlugin {
         MoEngagePluginBridge.sharedInstance.setPluginBridgeDelegate(self, payload: payload)
         MoEngagePluginBridge.sharedInstance.pluginInitialized(payload)
     }
+    
+    private func registerForProvisionalPush() {
+        if #available(iOS 12.0, *) {
+            MoEngagePluginBridge.sharedInstance.registerForProvisionalPush()
+        } else {
+            MoEngageLogger.logDefault(message: "Register for Provisional Push is not supported below iOS 12")
+        }
+    }
+    
 }