+++ packages/moengage_flutter/moengage_flutter_platform_interface/lib/src/model/inapp/self_handled_campaign.dart
@@ -1,7 +1,9 @@
+import 'display_rules.dart';
+
 /// Self Handled Inpp Campaign Data
 class SelfHandledCampaign {
   /// [SelfHandledCampaign] Constructor
-  SelfHandledCampaign(this.payload, this.dismissInterval);
+  SelfHandledCampaign(this.payload, this.dismissInterval, this.displayRules);
 
   /// Self handled campaign payload.
   String payload;
@@ -9,8 +11,11 @@ class SelfHandledCampaign {
   /// Interval after which in-app should be dismissed, unit - Seconds
   int dismissInterval;
 
+  /// InApp Campaign Display Rules
+  Rules displayRules;
+
   @override
   String toString() {
-    return '{\npayload: $payload\ndismissInterval: $dismissInterval\n}';
+    return 'SelfHandledCampaign{payload: $payload, dismissInterval: $dismissInterval, rules: $displayRules}';
   }
 }